move the NGCRegistryClient and S3Client to utils
